Avoid reporting serviceStatus of NotRunning whilst kamd is loading
Consumer::ServiceStatus is tri-state.
NotRunning, Unknown (loading), Running.

The backend was a simple boolean whether the service was up or not.
The library had code to try and start kactivitymanagerd on boot, or wait
until it is booted. However, this was not reflected in the service
status.

This leads to a race on plasma boot where the service status is reported
as "NotRunning" during initial startup.

A new enum is used as it's not a 1:1 match with Consumer::ServiceStatus,
when the backend is up, an async request of the activities is performed
and then the consumer is declared ready.

BUG: 466193