Initialise offline defaults on startup
A recent commit made activities library handle the case of connecting whilst the
daemon was booting more reliable.

In that change the call to loadOfflineDefaults was removed from the main path
when service status changes. It used to be on every service status changed. This
line was important as no other initialisation takes place.

We always want the null activity to be returned whilst real ones are not loaded.

Amends 9ac4e1db21f8ef1647023c24d91ac334d19b9b17